Simplify further.
